# docs/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "FastAPI",
		"version": "0.1.0"
	},
	"paths": {
...
		"/tasks/start": {
			"post": {
				"tags": [
					"tasks"
				],
				"summary": "Start a task",
				"description": "Start a task asynchronously and return its id.",
				"operationId": "start_task",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PipelineTask"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskStartResponse"
								}
							}
						}
					},
					"400": {
						"description": "Request validation failed",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"detail": {
											"type": "string"
										}
									},
									"type": "object"
								},
								"example": {
									"detail": "request validation failed"
								}
							}
						}
					},
					"500": {
						"description": "Task start failed",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"detail": {
											"type": "string"
										}
									},
									"type": "object"
								},
								"example": {
									"detail": "task start failed"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/tasks/running": {
			"get": {
				"tags": [
					"tasks"
				],
				"summary": "List running tasks",
				"description": "List active task runs.",
				"operationId": "list_running_tasks",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TaskRunningResponse"
									},
									"type": "array",
									"title": "Response List Running Tasks"
								}
							}
						}
					},
					"400": {
						"description": "Request validation failed",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"detail": {
											"type": "string"
										}
									},
									"type": "object"
								},
								"example": {
									"detail": "request validation failed"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"detail": {
											"type": "string"
										}
									},
									"type": "object"
								},
								"example": {
									"detail": "internal server error"
								}
							}
						}
					}
				}
			}
		},
		"/tasks/list": {
			"get": {
				"tags": [
					"tasks"
				],
				"summary": "List tasks",
				"description": "List tasks with filters and pagination.",
				"operationId": "list_tasks",
				"parameters": [
					{
						"name": "task_id",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "integer"
								},
								{
									"type": "null"
								}
							],
							"title": "Task Id"
						}
					},
					{
						"name": "spec",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "Spec"
						}
					},
					{
						"name": "created_at_from",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "Created At From"
						}
					},
					{
						"name": "created_at_to",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "Created At To"
						}
					},
					{
						"name": "started_at_from",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "Started At From"
						}
					},
					{
						"name": "started_at_to",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "Started At To"
						}
					},
					{
						"name": "finished_at_from",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "Finished At From"
						}
					},
					{
						"name": "finished_at_to",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "Finished At To"
						}
					},
					{
						"name": "last_heartbeat_at_from",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "Last Heartbeat At From"
						}
					},
					{
						"name": "last_heartbeat_at_to",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "Last Heartbeat At To"
						}
					},
					{
						"name": "page",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "integer"
								},
								{
									"type": "null"
								}
							],
							"title": "Page"
						}
					},
					{
						"name": "page_size",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "integer"
								},
								{
									"type": "null"
								}
							],
							"title": "Page Size"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"anyOf": [
									{
										"type": "array",
										"items": {
											"$ref": "#/components/schemas/TaskState"
										}
									},
									{
										"type": "null"
									}
								],
								"title": "State"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskListResponse"
								}
							}
						}
					},
					"400": {
						"description": "Request validation failed",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"detail": {
											"type": "string"
										}
									}
								},
								"example": {
									"detail": "request validation failed"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"detail": {
											"type": "string"
										}
									}
								},
								"example": {
									"detail": "internal server error"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/tasks/{task_id}": {
			"get": {
				"tags": [
					"tasks"
				],
				"summary": "Get task status",
				"description": "Get task status by id.",
				"operationId": "get_task_status",
				"parameters": [
					{
						"name": "task_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"title": "Task Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskStatusResponse"
								}
							}
						}
					},
					"400": {
						"description": "Request validation failed",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"detail": {
											"type": "string"
										}
									}
								},
								"example": {
									"detail": "request validation failed"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"detail": {
											"type": "string"
										}
									}
								},
								"example": {
									"detail": "internal server error"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/tasks/cancel/{task_id}": {
			"post": {
				"tags": [
					"tasks"
				],
				"summary": "Cancel a task",
				"description": "Cancel a pending or running task by id.",
				"operationId": "cancel_task",
				"parameters": [
					{
						"name": "task_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"title": "Task Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskStatusResponse"
								}
							}
						}
					},
					"400": {
						"description": "Request validation failed",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"detail": {
											"type": "string"
										}
									}
								},
								"example": {
									"detail": "request validation failed"
								}
							}
						}
					},
					"500": {
						"description": "Task cancel failed",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"detail": {
											"type": "string"
										}
									}
								},
								"example": {
									"detail": "task cancel failed"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"Arguments": {
				"properties": {
					"params": {
						"additionalProperties": true,
						"type": "object",
						"title": "Params"
					}
				},
				"type": "object",
				"title": "Arguments",
				"description": "Lightweight, JSON-serializable task arguments."
			},
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
...
			"PipelineTask": {
				"properties": {
					"spec": {
						"$ref": "#/components/schemas/TaskSpec"
					},
					"pipeline_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Pipeline Id"
					},
					"source": {
						"type": "string",
						"title": "Source"
					},
					"task_type": {
						"type": "string",
						"title": "Task Type"
					},
					"arguments": {
						"$ref": "#/components/schemas/Arguments"
					},
					"options": {
						"additionalProperties": true,
						"type": "object",
						"title": "Options"
					}
				},
				"type": "object",
				"required": [
					"spec",
					"source",
					"task_type"
				],
				"title": "PipelineTask",
				"description": "Task payload for pipeline execution."
			},
...
			"TaskListItem": {
				"properties": {
					"task_id": {
						"type": "integer",
						"title": "Task Id"
					},
					"idempotency_key": {
						"type": "string",
						"title": "Idempotency Key"
					},
					"spec": {
						"type": "string",
						"title": "Spec"
					},
					"state": {
						"$ref": "#/components/schemas/TaskState"
					},
					"attempt": {
						"type": "integer",
						"title": "Attempt"
					},
					"progress": {
						"type": "string",
						"title": "Progress"
					},
					"error": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Error"
					},
					"created_at": {
						"type": "string",
						"title": "Created At"
					},
					"started_at": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Started At"
					},
					"finished_at": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Finished At"
					},
					"last_heartbeat_at": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Last Heartbeat At"
					},
					"task_payload": {
						"additionalProperties": true,
						"type": "object",
						"title": "Task Payload"
					}
				},
				"type": "object",
				"required": [
					"task_id",
					"idempotency_key",
					"spec",
					"state",
					"attempt",
					"progress",
					"error",
					"created_at",
					"started_at",
					"finished_at",
					"last_heartbeat_at",
					"task_payload"
				],
				"title": "TaskListItem",
				"description": "Task list item including task payload."
			},
			"TaskListMeta": {
				"properties": {
					"page": {
						"type": "integer",
						"title": "Page"
					},
					"page_size": {
						"type": "integer",
						"title": "Page Size"
					},
					"total": {
						"type": "integer",
						"title": "Total"
					}
				},
				"type": "object",
				"required": [
					"page",
					"page_size",
					"total"
				],
				"title": "TaskListMeta",
				"description": "Metadata for task list response."
			},
			"TaskListResponse": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/TaskListItem"
						},
						"type": "array",
						"title": "Items"
					},
					"meta": {
						"$ref": "#/components/schemas/TaskListMeta"
					}
				},
				"type": "object",
				"required": [
					"items",
					"meta"
				],
				"title": "TaskListResponse",
				"description": "Response payload for task list endpoint."
			},
			"TaskRunningResponse": {
				"properties": {
					"task_id": {
						"type": "integer",
						"title": "Task Id"
					},
					"state": {
						"$ref": "#/components/schemas/TaskState"
					},
					"progress": {
						"type": "string",
						"title": "Progress"
					},
					"started_at": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Started At"
					},
					"error": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Error"
					}
				},
				"type": "object",
				"required": [
					"task_id",
					"state",
					"progress",
					"started_at",
					"error"
				],
				"title": "TaskRunningResponse",
				"description": "Response payload for running tasks."
			},
			"TaskSpec": {
				"type": "string",
				"enum": [
					"noop_sleep",
					"pipeline",
					"get_stock_info",
					"get_stock_hist_unadj",
					"get_fundamental_data",
					"get_adj_factor",
					"get_index_info",
					"get_index_hist_stock",
					"get_index_hist_bond",
					"get_index_hist_gold",
					"get_index_hist_global",
					"get_internal_index",
					"get_fund_info",
					"get_fund_hist_index",
					"get_fund_hist_money",
					"get_etf_info",
					"get_etf_hist",
					"get_market_factors",
					"get_gold_cftc_report",
					"get_gold_future_curve",
					"get_fund_beta",
					"get_rt_stock_hist_unadj",
					"get_rt_index_hist",
					"get_rt_etf_hist",
					"get_rt_market_factors"
				],
				"title": "TaskSpec",
				"description": "Enumerated task specifications used by the task runtime."
			},
			"TaskStartResponse": {
				"properties": {
					"task_id": {
						"type": "integer",
						"title": "Task Id"
					},
					"state": {
						"$ref": "#/components/schemas/TaskState"
					},
					"idempotency_key": {
						"type": "string",
						"title": "Idempotency Key"
					}
				},
				"type": "object",
				"required": [
					"task_id",
					"state",
					"idempotency_key"
				],
				"title": "TaskStartResponse",
				"description": "Response payload for task start requests."
			},
			"TaskState": {
				"type": "string",
				"enum": [
					"PENDING",
					"RUNNING",
					"SUCCEEDED",
					"FAILED",
					"CANCELLED"
				],
				"title": "TaskState",
				"description": "Lifecycle states for background tasks."
			},
			"TaskStatusResponse": {
				"properties": {
					"task_id": {
						"type": "integer",
						"title": "Task Id"
					},
					"idempotency_key": {
						"type": "string",
						"title": "Idempotency Key"
					},
					"spec": {
						"type": "string",
						"title": "Spec"
					},
					"state": {
						"$ref": "#/components/schemas/TaskState"
					},
					"attempt": {
						"type": "integer",
						"title": "Attempt"
					},
					"progress": {
						"type": "string",
						"title": "Progress"
					},
					"error": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Error"
					},
					"created_at": {
						"type": "string",
						"title": "Created At"
					},
					"started_at": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Started At"
					},
					"finished_at": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Finished At"
					},
					"last_heartbeat_at": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Last Heartbeat At"
					}
				},
				"type": "object",
				"required": [
					"task_id",
					"idempotency_key",
					"spec",
					"state",
					"attempt",
					"progress",
					"error",
					"created_at",
					"started_at",
					"finished_at",
					"last_heartbeat_at"
				],
				"title": "TaskStatusResponse",
				"description": "Response payload for task status lookups."
			},
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					}
				},
				"type": "object",
				"required": [
					"loc",
					"msg",
					"type"
				],
				"title": "ValidationError"
			}
		}
	}
}